+++ src/Resources/contao/dca/tl_synapsis_subscription.php
@@ -8,8 +8,8 @@ declare(strict_types=1);
  * @license LGPL-3.0-or-later
  */
 
-use Contao\DataContainer;
 use Contao\DC_Table;
+use Schachbulle\ContaoSynapsisBundle\SchachbulleContaoSynapsisBundle;
 
 /*
  * Tabelle tl_synapsis_subscription
@@ -22,7 +22,7 @@ $GLOBALS['TL_DCA']['tl_synapsis_subscription'] = array
 (
     'config' => array
     (
-        'dataContainer' => method_exists(DataContainer::class, 'getDriverForTable') ? DC_Table::class : 'Table',
+        'dataContainer' => SchachbulleContaoSynapsisBundle::isContao5() ? DC_Table::class : 'Table',
         'notEditable'   => true,
         'closed'        => true,
         'sql' => array